重构form 使期能够提供signature,policy给前端使用
