+++ lib/upyun.rb
@@ -1,8 +1,8 @@
 require 'upyun/version'
 require 'upyun/utils'
 require 'upyun/rest'
+require 'upyun/form_base'
 require 'upyun/form'
-require 'upyun/form_support'
 
 module Upyun
   DOMAIN = 'api.upyun.com'
